requirements.txt
brawlstats
discord
python-dotenv